Jdelete methods updated - WIP

+++ apiUrnaEletronica/src/main/java/com/adelmo/apiUrnaEletronica/services/CandidateService.java
@@ -1,8 +1,10 @@
 package com.adelmo.apiUrnaEletronica.services;
 
 import com.adelmo.apiUrnaEletronica.entities.Candidate;
+import com.adelmo.apiUrnaEletronica.exceptions.ElectionExceptions;
 import com.adelmo.apiUrnaEletronica.repositories.CandidateRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -15,7 +17,11 @@ public class CandidateService {
     private CandidateRepository candidateRepository;
 
     public Candidate createCandidate (Candidate candidate) {
-        return candidateRepository.save(candidate);
+        try {
+            return candidateRepository.save(candidate);
+        } catch (DataIntegrityViolationException exception) {
+            throw new ElectionExceptions("Candidate with the name, " + candidate.getName() + ", is already registered for the election.");
+        }
     }
 
     public Optional<Candidate> findCandidateById (Long id) {
@@ -27,7 +33,14 @@ public class CandidateService {
     }
 
     public void deleteCandidate (Long id) {
-        candidateRepository.deleteById(id);
+        Optional<Candidate> candidate = candidateRepository.findById(id);
+        if(candidate.isPresent()) {
+            if (candidate.get().getReceivedVotes() == 0) {
+                candidateRepository.deleteById(id);
+            } else {
+                throw new ElectionExceptions(candidate.get().getName()+" already has a score calculated and can't be deleted.");
+            }
+        }
     }
 
     public void countingReceivedVotes (Candidate candidate) {
